Win build fixes
Getting boinccas to build required some stuff I don't understand
