+++ samples/vboxwrapper/vbox_vboxmanage.cpp
@@ -46,6 +46,7 @@ using std::string;
 #include "str_util.h"
 #include "str_replace.h"
 #include "util.h"
+#include "proc_control.h"
 #include "error_numbers.h"
 #include "procinfo.h"
 #include "network.h"
@@ -1392,7 +1393,11 @@ namespace vboxmanage {
                 vboxlog_msg("VM did not stop when requested.");
 
                 // Attempt to terminate the VM
-                retval = kill_program(vm_pid);
+#ifdef _WIN32
+                retval = kill_process(vm_pid_handle);
+#else
+                retval = kill_process(vm_pid);
+#endif
                 if (retval) {
                     vboxlog_msg("VM was NOT successfully terminated.");
                 } else {
@@ -1435,7 +1440,11 @@ namespace vboxmanage {
                 vboxlog_msg("VM did not power off when requested.");
 
                 // Attempt to terminate the VM
-                retval = kill_program(vm_pid);
+#ifdef _WIN32
+                retval = kill_process(vm_pid_handle);
+#else
+                retval = kill_process(vm_pid);
+#endif
                 if (retval) {
                     vboxlog_msg("VM was NOT successfully terminated.");
                 } else {
